rxsplit {whisker}	R Documentation

Split a character in three parts

Description

It differs from strsplit in that it only splits on the first occurrence and returns all parts of the string given

Usage

  rxsplit(x, pattern)

Arguments

x
character text to be split

pattern
pattern used for splitting

[Package whisker version 0.3-2 Index]