whisker.render {whisker}	R Documentation

Logicless templating

Description

Logicless templating

Usage

  whisker.render(template, data = parent.frame(),
    partials = list(), debug = FALSE)

Arguments

template
character with template text

data
named list or environment with variables that will be used during rendering

partials
named list with partial templates, will be used during template construction

debug
Used for debugging purposes, likely to disappear

Value

character with rendered template

Examples

template <- "Hello {{place}}!"
place <- "World"

whisker.render(template)

[Package whisker version 0.3-2 Index]